refactor: remove bitcoind service
